+++ agp-compatibility/agp-compatibility-7-3/build.gradle
@@ -1,0 +1,9 @@
+plugins {
+    id "java-library"
+    id 'org.jetbrains.kotlin.jvm'
+}
+
+dependencies {
+    implementation project(":agp-compatibility:agp-compatibility-api")
+    compileOnly "com.android.tools.build:gradle:7.3.0"
+}
