+++ agp-compatibility/agp-compatibility-7-3/src/main/java/com/likethesalad/tools/agpcompat73/strategies/PluginStrategy73.kt
@@ -1,15 +1,16 @@
-package com.likethesalad.tools.android.plugin.data.strategies.impl
+package com.likethesalad.tools.agpcompat73.strategies
 
 import com.android.build.api.dsl.CommonExtension
 import com.android.build.api.variant.AndroidComponentsExtension
 import com.android.build.api.variant.Variant
-import com.likethesalad.tools.android.plugin.data.impl.componentextension.ComponentAndroidExtension
-import com.likethesalad.tools.android.plugin.data.impl.componentextension.ComponentAndroidVariantData
-import com.likethesalad.tools.android.plugin.data.strategies.PluginStrategy
-import com.likethesalad.tools.android.plugin.extension.observable.VariantPublisher
+import com.likethesalad.tools.agpcompat.api.PluginStrategy
+import com.likethesalad.tools.agpcompat.api.bridges.AndroidExtension
+import com.likethesalad.tools.agpcompat.api.observable.VariantPublisher
+import com.likethesalad.tools.agpcompat73.bridges.AndroidExtension73
+import com.likethesalad.tools.agpcompat73.bridges.AndroidVariantData73
 import org.gradle.api.Project
 
-class DefaultPluginStrategy : PluginStrategy {
+class PluginStrategy73 : PluginStrategy {
 
     override fun configure(project: Project, publisher: VariantPublisher): AndroidExtension {
         val androidExtension = project.extensions.getByType(CommonExtension::class.java)
@@ -19,11 +20,11 @@ class DefaultPluginStrategy : PluginStrategy {
             addVariantData(it, publisher)
         }
 
-        return ComponentAndroidExtension(androidExtension)
+        return AndroidExtension73(androidExtension)
     }
 
     private fun addVariantData(variant: Variant, publisher: VariantPublisher) {
-        val androidVariantData = ComponentAndroidVariantData(variant)
+        val androidVariantData = AndroidVariantData73(variant)
         publisher.publish(androidVariantData)
     }
 }
