+++ agp-compatibility/agp-compatibility-7-3/src/main/resources/META-INF/services/com.likethesalad.tools.agpcompat.api.AgpCompatibilityEntrypoint
@@ -1,0 +1,1 @@
+com.likethesalad.tools.agpcompat73.AgpCompatibilityEntrypoint73
